+++ packages/houdini/src/codegen/generators/artifacts/tests/artifacts.test.ts
@@ -648,6 +648,11 @@ test('paginate over unions', async function () {
 		                                                            "visible": true
 		                                                        },
 
+		                                                        "id": {
+		                                                            "type": "ID",
+		                                                            "keyRaw": "id"
+		                                                        },
+
 		                                                        "__typename": {
 		                                                            "type": "String",
 		                                                            "keyRaw": "__typename",
@@ -835,6 +840,11 @@ test('overlapping query and fragment nested selection', async function () {
 		                                    "visible": true
 		                                },
 
+		                                "id": {
+		                                    "type": "ID",
+		                                    "keyRaw": "id"
+		                                },
+
 		                                "friends": {
 		                                    "type": "User",
 		                                    "keyRaw": "friends",
@@ -1000,6 +1010,11 @@ test('selections with interfaces', async function () {
 		                                    "visible": true
 		                                },
 
+		                                "aka": {
+		                                    "type": "String",
+		                                    "keyRaw": "aka"
+		                                },
+
 		                                "__typename": {
 		                                    "type": "String",
 		                                    "keyRaw": "__typename",
@@ -1144,6 +1159,11 @@ test('selections with unions', async function () {
 		                                    "type": "String",
 		                                    "keyRaw": "aka",
 		                                    "visible": true
+		                                },
+
+		                                "aka": {
+		                                    "type": "String",
+		                                    "keyRaw": "aka"
 		                                },
 
 		                                "__typename": {
@@ -1298,6 +1318,11 @@ test('selections with overlapping unions', async function () {
 		                                    "type": "String",
 		                                    "keyRaw": "aka",
 		                                    "visible": true
+		                                },
+
+		                                "aka": {
+		                                    "type": "String",
+		                                    "keyRaw": "aka"
 		                                },
 
 		                                "__typename": {
@@ -1472,6 +1497,11 @@ test('selections with unions of abstract types', async function () {
 		                                    "visible": true
 		                                },
 
+		                                "aka": {
+		                                    "type": "String",
+		                                    "keyRaw": "aka"
+		                                },
+
 		                                "__typename": {
 		                                    "type": "String",
 		                                    "keyRaw": "__typename",
@@ -1599,6 +1629,11 @@ test('selections with concrete types matching multiple abstract types', async fu
 		                                    "type": "String",
 		                                    "keyRaw": "name",
 		                                    "visible": true
+		                                },
+
+		                                "name": {
+		                                    "type": "String",
+		                                    "keyRaw": "name"
 		                                },
 
 		                                "cats": {