+++ packages/houdini-react/CHANGELOG.md
@@ -1,5 +1,26 @@
 # houdini-react
 
+## 1.0.7
+
+### Dependency Changes
+
+-   Updated dependencies [[`52326b5`](https://github.com/HoudiniGraphql/houdini/commit/52326b5b54c1e722d398031e4b61281379cb8820)]:
+    -   houdini@1.0.7
+
+## 1.0.6
+
+### Dependency Changes
+
+-   Updated dependencies [[`8fd052c`](https://github.com/HoudiniGraphql/houdini/commit/8fd052c1d59fbb37e17da1bc42ae386a660440ed)]:
+    -   houdini@1.0.6
+
+## 1.0.5
+
+### Dependency Changes
+
+-   Updated dependencies [[`f69f9f1`](https://github.com/HoudiniGraphql/houdini/commit/f69f9f1b12cf9bca5d0112db2e78c4d4e94b4845), [`92c533e`](https://github.com/HoudiniGraphql/houdini/commit/92c533e2ba0aae7ceaebe7407691ff36482a71f4)]:
+    -   houdini@1.0.5
+
 ## 1.0.4
 
 ### Dependency Changes