+++ packages/houdini/CHANGELOG.md
@@ -1,5 +1,25 @@
 # houdini
 
+## 1.0.7
+
+### 🐛 Fixes
+
+-   [#950](https://github.com/HoudiniGraphql/houdini/pull/950) [`52326b5`](https://github.com/HoudiniGraphql/houdini/commit/52326b5b54c1e722d398031e4b61281379cb8820) Thanks [@SeppahBaws](https://github.com/SeppahBaws)! - @include and @skip now add "undefined" to the generated TypeScript type
+
+## 1.0.6
+
+### 🦋 Feature/Fix
+
+-   [#948](https://github.com/HoudiniGraphql/houdini/pull/948) [`8fd052c`](https://github.com/HoudiniGraphql/houdini/commit/8fd052c1d59fbb37e17da1bc42ae386a660440ed) Thanks [@AlecAivazis](https://github.com/AlecAivazis)! - Add id to the generated selections of inline fragments
+
+## 1.0.5
+
+### 🐛 Fixes
+
+-   [#940](https://github.com/HoudiniGraphql/houdini/pull/940) [`f69f9f1`](https://github.com/HoudiniGraphql/houdini/commit/f69f9f1b12cf9bca5d0112db2e78c4d4e94b4845) Thanks [@AlecAivazis](https://github.com/AlecAivazis)! - Fix bug when relistening on a subscription
+
+-   [#947](https://github.com/HoudiniGraphql/houdini/pull/947) [`92c533e`](https://github.com/HoudiniGraphql/houdini/commit/92c533e2ba0aae7ceaebe7407691ff36482a71f4) Thanks [@AlecAivazis](https://github.com/AlecAivazis)! - Fix behavior for CacheAndNetwork policies
+
 ## 1.0.4
 
 ### ✨ Features